Indexer functioning ok
Changed to sets since JPA works better with them instead of lists
Changed to PostgresSQL database
Changed model
Changed jobs to be more simple

+++ ductr-imdb/src/main/java/com/ductr/ductrimdb/mapper/RatingMapper.java
@@ -16,8 +16,8 @@ public class RatingMapper implements FieldSetMapper<Title> {
 
     Title title = new Title();
     title.setTconst(fieldSet.readString("tconst"));
-    title.setRating(fieldSet.readDouble("averageRating"));
+    title.setRating(fieldSet.readFloat("averageRating"));
     return title;
   }
-  
+
 }
